refactor: split returned-parcel tests into their own file (issue 299)
posta-uncollected-run.integration.test.ts crossed eslint max-lines: 400
after adding the vratena_zasielka tests — extracted into
posta-uncollected-run-returned.integration.test.ts, same pattern as the
existing orders-http.integration.test.ts / orders-http-state split.
